src/requirements/base.txt
python-dotenv==1.0.0
kombu==5.2.4
